+++ webapi/src/test/scala/org/knora/webapi/e2e/v1/ResourcesV1R2RSpec.scala
@@ -14,7 +14,6 @@ import org.knora.webapi._
 import dsp.errors.AssertionException
 import dsp.errors.InvalidApiJsonException
 import dsp.errors.NotFoundException
-import dsp.errors.TriplestoreResponseException
 import org.knora.webapi.http.directives.DSPApiDirectives
 import org.knora.webapi.messages.OntologyConstants
 import org.knora.webapi.messages.store.triplestoremessages._
@@ -28,6 +27,7 @@ import org.knora.webapi.routing.v2.ResourcesRouteV2
 import org.knora.webapi.sharedtestdata.SharedOntologyTestDataADM._
 import org.knora.webapi.sharedtestdata.SharedTestDataADM
 import org.knora.webapi.sharedtestdata.SharedTestDataADM._
+import org.knora.webapi.store.triplestore.errors.TriplestoreResponseException
 import org.knora.webapi.util.AkkaHttpUtils
 import org.knora.webapi.util.MutableTestIri
 import org.scalatest.Assertion
@@ -284,7 +284,7 @@ class ResourcesV1R2RSpec extends R2RSpec {
       Get(
         "/v1/resources.html/http%3A%2F%2Frdfh.ch%2F0803%2Fc5058f3a?noresedit=true&reqtype=properties"
       ) ~> resourcesPathV1 ~> check {
-        //log.debug("==>> " + responseAs[String])
+        // log.debug("==>> " + responseAs[String])
         assert(status === StatusCodes.OK)
         assert(responseAs[String] contains "Physical description")
         assert(responseAs[String] contains "Location")
@@ -300,7 +300,7 @@ class ResourcesV1R2RSpec extends R2RSpec {
       Get(
         "/v1/resources.html/http%3A%2F%2Frdfh.ch%2F0803%2Fde6c38ce3401?noresedit=true&reqtype=properties"
       ) ~> resourcesPathV1 ~> check {
-        //log.debug("==>> " + responseAs[String])
+        // log.debug("==>> " + responseAs[String])
         assert(status === StatusCodes.OK)
         assert(responseAs[String] contains "preview")
         assert(responseAs[String] contains "Original filename")
@@ -828,7 +828,7 @@ class ResourcesV1R2RSpec extends R2RSpec {
         BasicHttpCredentials(anythingUserEmail, password)
       ) ~> resourcesPathV1 ~> check {
 
-        //println(response)
+        // println(response)
 
         // the route should reject the request because an IRI is wrong (formally valid though)
         assert(status == StatusCodes.NotFound, response.toString)
@@ -1837,16 +1837,18 @@ class ResourcesV1R2RSpec extends R2RSpec {
         maybeAppendValue(
           random = random,
           xmlStringBuilder = xmlStringBuilder,
-          value = s"""
-                     |<p0001-anything:hasText knoraType="richtext_value">This is a test in thing $i.</p0001-anything:hasText>
+          value =
+            s"""
+               |<p0001-anything:hasText knoraType="richtext_value">This is a test in thing $i.</p0001-anything:hasText>
                         """.stripMargin
         )
 
         maybeAppendValue(
           random = random,
           xmlStringBuilder = xmlStringBuilder,
-          value = s"""
-                     |<p0001-anything:hasTimeStamp knoraType="time_value">2019-08-28T15:13:10.968318Z</p0001-anything:hasTimeStamp>
+          value =
+            s"""
+               |<p0001-anything:hasTimeStamp knoraType="time_value">2019-08-28T15:13:10.968318Z</p0001-anything:hasTimeStamp>
                         """.stripMargin
         )