+++ webapi/src/test/scala/org/knora/webapi/models/OntologyModels.scala
@@ -104,7 +104,7 @@ object CreatePropertyRequest {
                                    |      "@id" : "$ontologyName:$subjectName"
                                    |    },
                                    |""".stripMargin
-      case None              => ""
+      case None => ""
     }
 
     val value = s"""{
@@ -196,7 +196,7 @@ object AddCardinalitiesRequest {
   ): AddCardinalitiesRequest = {
     val ontologyId                 = s"http://0.0.0.0:3333/ontology/0001/$ontologyName/v2"
     val restrictionsString: String = stringifyRestrictions(restrictions)
-    val value                      = s"""
+    val value = s"""
                    |{
                    |  "@id" : "$ontologyId",
                    |  "@type" : "owl:Ontology",