+++ webapi/src/test/scala/org/knora/webapi/store/cache/config/RedisTestConfig.scala
@@ -6,11 +6,10 @@ import zio._
 
 object RedisTestConfig {
   val hardcoded: ULayer[RedisConfig] = ZLayer.succeed(RedisConfig("localhost", 6379))
-  val redisTestContainer: ZLayer[RedisTestContainer, Nothing, RedisConfig] = {
+  val redisTestContainer: ZLayer[RedisTestContainer, Nothing, RedisConfig] =
     ZLayer {
       for {
         rtc <- ZIO.service[RedisTestContainer]
       } yield RedisConfig("localhost", rtc.container.getFirstMappedPort())
     }.tap(_ => ZIO.debug(">>> Redis Test Config Initialized <<<"))
-  }
 }