+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+### [23.0.3](https://www.github.com/dasch-swiss/dsp-api/compare/v23.0.2...v23.0.3) (2022-08-02)
+
+
+### Bug Fixes
+
+* **triplestore-connector:** stack crashes on invalid search (DEV-1154) ([#2140](https://www.github.com/dasch-swiss/dsp-api/issues/2140)) ([e5426dc](https://www.github.com/dasch-swiss/dsp-api/commit/e5426dca0888ca7ab064cbe292904217dbe8b17f))
+
+
+### Maintenance
+
+* **dependencies:** update akka-http-cors to 1.1.3 ([#2103](https://www.github.com/dasch-swiss/dsp-api/issues/2103)) ([5d0d522](https://www.github.com/dasch-swiss/dsp-api/commit/5d0d52267ce9e702215c7b2a6527df7e0eb822ef))
+* **dependencies:** update jwt-spray-json to 9.0.2 ([#2111](https://www.github.com/dasch-swiss/dsp-api/issues/2111)) ([6e54443](https://www.github.com/dasch-swiss/dsp-api/commit/6e54443e4c2234e432d54ca25b7ecbd3a8656a2f))
+* **dependencies:** update Saxon-HE to 11.4 ([#2137](https://www.github.com/dasch-swiss/dsp-api/issues/2137)) ([08c9f68](https://www.github.com/dasch-swiss/dsp-api/commit/08c9f68d427b230327aea149192c907cc3387092))
+* **dependencies:** update scalatest to 3.2.13 ([#2138](https://www.github.com/dasch-swiss/dsp-api/issues/2138)) ([a345079](https://www.github.com/dasch-swiss/dsp-api/commit/a345079999de24cabed8235f055bab46cdf6094b))
+* **dependencies:** update spring-security-core to 5.6.6 ([#2130](https://www.github.com/dasch-swiss/dsp-api/issues/2130)) ([c83645d](https://www.github.com/dasch-swiss/dsp-api/commit/c83645d6c4a4d949d40c0ec1507106c6071434c2))
+* **dependencies:** update spring-security-core to 5.7.2 ([#2139](https://www.github.com/dasch-swiss/dsp-api/issues/2139)) ([3a12562](https://www.github.com/dasch-swiss/dsp-api/commit/3a1256216022770eca0515b21c8ea663ba6b3425))
+* **dependencies:** update titanium-json-ld to 1.3.1 ([#2104](https://www.github.com/dasch-swiss/dsp-api/issues/2104)) ([4850525](https://www.github.com/dasch-swiss/dsp-api/commit/4850525494aa2f949b7dfaa2fb1be6f7ff14b684))
+
+### [23.0.2](https://www.github.com/dasch-swiss/dsp-api/compare/v23.0.1...v23.0.2) (2022-07-29)
+
+
+### Bug Fixes
+
+* **ontology:** link value property is still not editable after updating the property metadata (DEV-1116) ([#2133](https://www.github.com/dasch-swiss/dsp-api/issues/2133)) ([d5b48db](https://www.github.com/dasch-swiss/dsp-api/commit/d5b48db1528d3a8f0c0d3f08bc97564d01e383ce))
+* **sipi:** cookie parsing can cause an error which leads to 404 for images (DEV-1135) ([#2134](https://www.github.com/dasch-swiss/dsp-api/issues/2134)) ([bd023a5](https://www.github.com/dasch-swiss/dsp-api/commit/bd023a5836c1c741f094f9b6c974adc54a8d798d))
+
+
+### Maintenance
+
+* add dependency checking ([#2100](https://www.github.com/dasch-swiss/dsp-api/issues/2100)) ([8017b1f](https://www.github.com/dasch-swiss/dsp-api/commit/8017b1fe96e0728d8d3b112c25b121aa7af073c2))
+* add dependency checking ([#2102](https://www.github.com/dasch-swiss/dsp-api/issues/2102)) ([856277b](https://www.github.com/dasch-swiss/dsp-api/commit/856277b0c7d843f4ff563bf8350e99c579073d14))
+* Improve validation of GUI elements and GUI attributes (DEV-1082) ([#2098](https://www.github.com/dasch-swiss/dsp-api/issues/2098)) ([5cec8ba](https://www.github.com/dasch-swiss/dsp-api/commit/5cec8bac0b73237c85a72a33d2e3cd20a7cebee1))
+* **v3:** add role slice (DEV-1010) ([#2099](https://www.github.com/dasch-swiss/dsp-api/issues/2099)) ([6920716](https://www.github.com/dasch-swiss/dsp-api/commit/6920716e72b9acb32f5444bc6848bfda2986fb93))
+
 ### [23.0.1](https://www.github.com/dasch-swiss/dsp-api/compare/v23.0.0...v23.0.1) (2022-07-19)